requirements.txt
astroid==2.11.5
asttokens==2.0.5
attrs==21.4.0
autopep8==1.6.0
backcall==0.2.0
debugpy==1.6.0
decorator==5.1.1
dill==0.3.5.1
entrypoints==0.4
executing==0.8.3
iniconfig==1.1.1
ipykernel==6.13.1
ipython==8.4.0
isort==5.10.1
jedi==0.18.1
joblib==1.1.0
jupyter-client==7.3.2
jupyter-core==4.10.0
lazy-object-proxy==1.7.1
matplotlib-inline==0.1.3
mccabe==0.7.0
MedPy==0.4.0
nest-asyncio==1.5.5
numpy==1.22.4
packaging==21.3
pandas==1.4.2
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.1.1
platformdirs==2.5.2
pluggy==1.0.0
prompt-toolkit==3.0.29
psutil==5.9.1
ptyprocess==0.7.0
pure-eval==0.2.2
py==1.11.0
pycodestyle==2.8.0
Pygments==2.12.0
pylint==2.13.9
pyparsing==3.0.9
pytest==7.1.2
python-dateutil==2.8.2
pytz==2022.1
pyzmq==23.1.0
scikit-learn==1.1.1
scipy==1.8.1
SimpleITK==2.1.1.2
six==1.16.0
sklearn==0.0
stack-data==0.2.0
threadpoolctl==3.1.0
toml==0.10.2
tomli==2.0.1
torch==1.10.1+cu111
torchvision==0.11.2+cu111
tornado==6.1
traitlets==5.2.2.post1
typing-extensions==4.2.0
wcwidth==0.2.5
wrapt==1.14.1
